### 2024-02-22 #160 @KatyBrown
plot_phylo
Submitting Author: Katy Brown (@KatyBrown)
All current maintainers: Katy Brown (@KatyBrown)
Package Name: plot_phylo
One-Line Description of Package: A Python package to plot a phylogenetic tree on an existing Matplotlib axis.
Repository Link:  https://github.com/KatyBrown/plot_phylo
Version submitted:   1.9.0
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13871592.svg)](https://doi.org/10.5281/zenodo.13871592)
Version accepted: 0.1.11

---

`plot_phylo` is a Python package which allows the user to plot phylogenetic trees onto an existing matplotlib axis. This means annotations can be added using matplotlib functionality and figures including phylogenies can be included in automated figure generation.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

Researchers working with phylogenetic tree data, a common data format for evolutionary and functional analysis of biological data.

plot_phylo depends on the functionality of [ETE Toolkit](http://etetoolkit.org/), a widely used visualisation tool. ETE provides excellent tree visualisations but these act as a stand-alone image, which can't be incorporated into another programmatically generated figure except as an uneditable embedded image. plot_phylo trees are subplots of open matplotlib figures, so custom annotations can be added, aspects of the image can be updated and other subplots can show additional data.
