### 2024-02-27 #160 @isabelizimm
---
---

Thank you again for submitting `plot_phylo` to pyOpenSci! I've gone through your package, and you've got a great start!

Blockers to beginning review process:

- Right now, this package lacks a documentation website. There is some documentation in [pyOpenSci's packaging guide](https://www.pyopensci.org/python-package-guide/documentation/hosting-tools/intro.html) on how to set up a site with Sphinx and publish it to GitHub. You have lovely docstrings already in place for `plot_phylo`, so you've already done a lot of the heavy lifting! This site should help with the maintenance of the package for you as well-- Sphinx will autogenerate API docs so you won't have to write out `Parameters` and such yourself 🙂

Some smaller nits:

- I would recommend updating testing in CI to more modern Python versions. It looks like [you are only testing](https://github.com/KatyBrown/plot_phylo/blob/cbecee6a37ee6835dc5bdc072568bf8c1e264cc2/.github/workflows/main.yml#L15C27-L15C36) on Python 3.7 and 3.8; [Python 3.7 has been end-of-lifed, and Python 3.8 will be end-of-life in Oct of this year](https://devguide.python.org/versions/).
- Your `__init__.py` and `_version.py` files have two different versions. It looks like you only use the `_version.py` file though.

If you have any questions or need any support with these tasks, pyOpenSci community would love to help out! The best places to get in touch are via [discourse](https://pyopensci.discourse.group/) or Slack--let me know if you need an invite. And, of course, you can also reply here (but then you will be restricted to just my opinions 😂)!
