### 2024-03-01 #160 @isabelizimm
Following up, @ctb will the editor leading the review for `plot_phylo` once the above items have been addressed! Titus also has a few things to finish up before starting this role, so I'll check back here in about 2 weeks to check in if both parties are ready to kick off this review 😄
