### 2024-03-28 f6e8036
adjusted alignment of lines and text
