### 2024-03-28 b071940
basic sphinx docs
