+++ man/pages/functions.rst
@@ -1,0 +1,7 @@
+Functions
+======================
+
+.. automodule:: plot_phylo
+   :members:
+   :undoc-members:
+   :show-inheritance:
