+++ man/pages/matplotlib.md
@@ -1,0 +1,1 @@
+# Incorporating Matplotlib Parameters
