+++ man/pages/installation.md
@@ -1,0 +1,18 @@
+## Installation
+
+**Requirements**
+
+* python >= 3.6
+* matplotlib >= 2.1.1
+* ete3 >= 3.1.0
+
+## Installation
+The module can be installed using pip
+
+`pip install plot_phylo`
+
+You can also download the latest release [here](https://github.com/KatyBrown/plot_phylo/releases/latest)
+
+Or clone the GitHub repository directly.
+
+`git clone git@github.com:KatyBrown/plot_phylo.git`
