+++ man/pages/detailed.md
@@ -1,0 +1,1 @@
+# Detailed Example
