+++ man/contents.rst
@@ -1,0 +1,12 @@
+plot_phylo
+==========
+
+.. toctree::
+   :maxdepth: 3
+
+   pages/introduction.md
+   pages/installation.md
+   pages/quickstart.md
+   pages/parameters.md
+   pages/functions.rst
+   pages/detailed.md
