+++ .readthedocs.yaml
@@ -17,7 +17,7 @@ build:
 
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
-  configuration: docs/conf.py
+  configuration: man/conf.py
 
 # Optionally build your docs in additional formats such as PDF and ePub
 # formats:
