+++ man/contents.rst
@@ -4,9 +4,11 @@ plot_phylo
 .. toctree::
    :maxdepth: 3
 
+   index.md
    pages/introduction.md
    pages/installation.md
    pages/quickstart.md
    pages/parameters.md
+   pages/matplotlib.md
    pages/functions.rst
    pages/detailed.md
