+++ man/index.md
@@ -1,0 +1,7 @@
+# plot_phylo
+1. [Installation](pages/installation.md)
+2. [Quick Start](pages/quickstart.md)
+3. [Incorporating Matplotlib Parameters](pages/matplotlib.md)
+4. [Parameters](pages/parameters.md)
+5. [Functions](pages/functions.md)
+6. [Detailed Example](pages/detailed.md)
