+++ man/pages/introduction.md
@@ -1,18 +1,0 @@
-# Introduction
-
-This module allows the user to plot a phylogenetic tree on an existing matplotlib axis.
-
-This means that:
-* Phylogenies can be incorporated into existing plots.
-* Annotations can be added using standard matplotlib functionality.
-* Plots can be output in png, pdf, svg or tiff formats.
-* Automatically generated and updated figures can include phylogenies
-
-1. [Installation](installation.md)
-2. [Quick Start](quickstart.md)
-3. [Incorporating Matplotlib Parameters](matplotlib.md)
-4. [Parameters](parameters.md)
-5. [Functions](functions.md)
-6. [Detailed Example](detailed.md)
-
-![Illustration](examples/layered.png "Illustration")
