+++ man/index.md
@@ -1,7 +1,18 @@
 # plot_phylo
+![Illustration](pages/examples/layered.png "Illustration")
+
+This module allows the user to plot a phylogenetic tree on an existing matplotlib axis.
+
+This means that:
+* Phylogenies can be incorporated into existing plots.
+* Annotations can be added using standard matplotlib functionality.
+* Plots can be output in png, pdf, svg or tiff formats.
+* Automatically generated and updated figures can include phylogenies
+
 1. [Installation](pages/installation.md)
 2. [Quick Start](pages/quickstart.md)
-3. [Incorporating Matplotlib Parameters](pages/matplotlib.md)
-4. [Parameters](pages/parameters.md)
+3. [Parameters](pages/parameters.md)
+4. [Incorporating Matplotlib Elements](pages/matplotlib.md)
 5. [Functions](pages/functions.md)
 6. [Detailed Example](pages/detailed.md)
+

[…]
