+++ .readthedocs.yaml
@@ -30,8 +30,3 @@ sphinx:
 python:
   install:
     - requirements: man/requirements.txt
-
-steps:
-  - command: |
-      pip install -r man/requirements.txt
-      sphinx-build -b html man/_build man
