+++ .readthedocs.yaml
@@ -15,10 +15,6 @@ build:
     # rust: "1.64"
     # golang: "1.19"
 
-  steps:
-    - command: |
-        pip install -r man/requirements.txt
-        sphinx-build -b html man/_build man
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: man/conf.py
