+++ man/conf.py
@@ -12,6 +12,7 @@
 #
 # import os
 import sys
+import os
 sys.path.insert(0, "../plot_phylo")
 sys.path.insert(0, os.path.abspath(".."))
 
