+++ man/conf.py
@@ -41,6 +41,8 @@ templates_path = ['_templates']
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
+source_suffix = ['.rst', '.md']
+
 root_doc = 'contents'
 
 # -- Options for HTML output -------------------------------------------------
