+++ man/functions.rst
@@ -1,7 +1,0 @@
-Functions
-======================
-
-.. automodule:: plot_phylo
-   :members:
-   :undoc-members:
-   :show-inheritance:
