### 2024-04-12 a3017aa
updated main.yml for requirements.txt
