+++ .github/workflows/main.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python3 -m pip install --upgrade pip
-          pip install -r requirements-test.txt
+          pip install -r requirements.txt
           pip install pytest
           pip3 install -e .
 
