+++ man/conf.py
@@ -15,7 +15,7 @@ import sys
 import os
 sys.path.insert(0, "../plot_phylo")
 sys.path.insert(0, os.path.abspath(".."))
-
+autodoc_mock_imports = ['ete3']
 # -- Project information -----------------------------------------------------
 
 project = 'plot_phylo'
