+++ plot_phylo/__init__.py
@@ -1,4 +1,3 @@
 from plot_phylo.plot_phylo import *
 
-__version__ = "0.1.8"
 __author__ = 'Katy Brown, Duncan Cross'
