+++ README.md
@@ -12,462 +12,22 @@ This means that:
 * Plots can be output in png, pdf, svg or tiff formats.
 * Automatically generated and updated figures can include phylogenies
 
-The module depends on the [ETE Toolkit](http://etetoolkit.org/), an existing Python framework for analysing and visualising phylogenetic trees, plus the [matplotlib](https://matplotlib.org/) visualisation library.
+Full documentation is available via [ReadTheDocs](https://plot-phylo.readthedocs.io/en/latest/index.html).
 
-The tree used in these examples is from the [10K trees](https://10ktrees.nunn-lab.org/) project and is illustrative only. The secondary tree shown in the detailed example is manually edited and not based on biological data.
+# Installation
 
-## Installation
+**Requirements**
+
+* python >= 3.6
+* matplotlib >= 2.1.1
+* ete3 >= 3.1.0
+
 The module can be installed using pip
 
 `pip install plot_phylo`
 
 You can also download the latest release [here](https://github.com/KatyBrown/plot_phylo/releases/latest)
 
-Or clone this repository directly.
+Or clone the GitHub repository directly.
 
-## Quick Start
-
-To draw a phylogeny under the default settings onto a blank figure.
-
-```
-import matplotlib.pyplot as plt
-import plot_phylo
-
-# Create an empty plot, 8in (width) by 10in (height) - matplotlib
-f = plt.figure(figsize=(8, 10))
-
-# Add an axis - matplotlib
-ax = plt.subplot()
-
-# Plot the tree on this axis, using the default settings - plot_phylo
-results = plot_phylo.plot_phylo("examples/primates.nw", ax)
-
-# Save the tree - matplotlib
-plt.savefig("examples/basic_plot.png", bbox_inches='tight')
-```
-
-The output of this function is:
-![Basic Tree](./examples/basic_plot.png "Basic Tree")
-
-## Adding matplotlib Elements
-One of the main advantages of this module is that, as plots are generated using matplotlib and added to an active axis, other elements of matplotlib plots can be incorporated - the plot remains activate and can be edited.
-
-The plot_phylo function returns a dictionary where the keys are the tip labels and the values are the boundaries of the text boxes containing the tip labels, in axis units. 
-
-For example, the first few items in the list returned from the function call above are as follows:
-
-
-```
-{'Cercopithecus neglectus': {'index': 0,
-  'xmin': 9.656,
-  'xmax': 16.336,
-  'ymin': 9.882,
-  'ymax': 10.118,
-  'ymid': 10.0,
-  'xmid': 12.883},
- 'Macaca mulatta': {'index': 1,
-  'xmin': 9.609,
-  'xmax': 14.2,
-  'ymin': 9.215,
-  'ymax': 9.452,
-  'ymid': 9.333,
-  'xmid': 12.101},
- 'Colobus angolensis': {'index': 2,
-  'xmin': 9.649,
-  'xmax': 15.016,
-  'ymin': 8.548,
-  'ymax': 8.785,
-  'ymid': 8.667,
-  'xmid': 12.883},
-...
-```
-
-`xmin`, `xmid` and `xmax` are the positions of the left, centre and right of the tip label text on the x-axis.
-
-`ymin`, `ymid` and `ymax` are the positions of the bottom, centre and top of the tip label text on the y-axis.
-
-`ind` is the position of this label on the tree, reading from top to bottom, starting at 0.
-
[…]
