### 2024-04-12 9142fbc
added additional python versions to testing
