+++ .github/workflows/main.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        python-version: [ 3.7, 3.8, 3.9, 3.10, 3.11, 3.12]
+        python-version: [ 3.7, 3.8, 3.9, '3.10', 3.11, 3.12 ]
       fail-fast: false
     # a job is a seq of steps
     steps:
