### 2024-04-12 #160 @KatyBrown
Hi @isabelizimm, very sorry for the long delay, I've now made the changes you asked for. The current version of the tool is now v0.1.9 - I've edited this in my original post.

Specifically I have:
* Extended the documentation and moved it to ReadTheDocs [here](https://plot-phylo.readthedocs.io/en/latest/index.html)
* Added testing in Python 3.9 to 3.12
* Corrected the mistake with the version

Please let me know if there's anything else I need to do, otherwise I think I'm now ready to proceed to the next step :)

Thank you!
