### 2024-05-03 a313b03
updated setup-python version in CI workflow
