### 2024-05-03 21fa260
to avoid this issue https://github.com/actions/runner-images/issues/9770 added this workaround issuecomment-2085623315
