### 2024-05-03 4947f33
fixed link to clone from github in install instructions
