### 2024-05-03 0490e84
updates to detailed example man page
