### 2024-07-01 #160 @ctb
## Editor response to review:

---

## Editor comments
:wave: Hi @ammaraziz and @jkanche! Thank you for volunteering to review
for pyOpenSci! Sorry it took me so long to post this, I blame ...umm, summer! Yeah, it's summer time! But seriously, thank you so much!!

- [ ] reviewer 1 survey completed (ammaraziz)
- [ ] reviewer 2 survey completed (jkanche)

Please get in touch with any questions or concerns! Your review is due: July 22nd, 2024!
---

Reviewers: @ammaraziz and @jkanche
Due date: July 22nd, 2024
