### 2024-08-05 #160 @jkanche
---

- Cannot run tests following the instructions

I cloned the repo and installed it in edit mode:

```pyt
⋊> ~/P/p/t/plot_phylo on main ◦ pytest --cov=plot_phylo                                   (reviews) 16:09:15
=========================================== test session starts ============================================
platform darwin -- Python 3.11.9, pytest-8.2.2, pluggy-1.5.0
rootdir: /Users/kancherj/Projects/public/trash/plot_phylo
plugins: cov-5.0.0
collected 0 items / 3 errors
/opt/homebrew/lib/python3.11/site-packages/coverage/control.py:888: CoverageWarning: No data was collected. (no-data-collected)
  self._warn("No data was collected.", slug="no-data-collected")

================================================== ERRORS ==================================================
______________________________ ERROR collecting tests/test_get_boxes_data.py _______________________________
ImportError while importing test module '/Users/kancherj/Projects/public/trash/plot_phylo/tests/test_get_boxes_data.py'.
/opt/homebrew/Cellar/python@3.11/3.11.9/Frameworks/Python.framework/Versions/3.11/lib/python3.11/importlib/__init__.py:126: in import_module
tests/test_get_boxes_data.py:2: in <module>
E   ModuleNotFoundError: No module named 'matplotlib'
________________________________ ERROR collecting tests/test_plot_phylo.py _________________________________
ImportError while importing test module '/Users/kancherj/Projects/public/trash/plot_phylo/tests/test_plot_phylo.py'.
/opt/homebrew/Cellar/python@3.11/3.11.9/Frameworks/Python.framework/Versions/3.11/lib/python3.11/importlib/__init__.py:126: in import_module
tests/test_plot_phylo.py:2: in <module>
E   ModuleNotFoundError: No module named 'matplotlib'
______________________________ ERROR collecting tests/test_plot_phylo_data.py ______________________________
ImportError while importing test module '/Users/kancherj/Projects/public/trash/plot_phylo/tests/test_plot_phylo_data.py'.
/opt/homebrew/Cellar/python@3.11/3.11.9/Frameworks/Python.framework/Versions/3.11/lib/python3.11/importlib/__init__.py:126: in import_module
tests/test_plot_phylo_data.py:2: in <module>
    import plot_phylo
E   ModuleNotFoundError: No module named 'plot_phylo'

---------- coverage: platform darwin, python 3.11.9-final-0 ----------
Name                       Stmts   Miss  Cover   Missing
--------------------------------------------------------
plot_phylo/plot_phylo.py     144    144     0%   2-586
--------------------------------------------------------

========================================= short test summary info ==========================================
ERROR tests/test_get_boxes_data.py
ERROR tests/test_plot_phylo.py
ERROR tests/test_plot_phylo_data.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 3 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
```
- Add a short description and link to docs for the repository metadata. Helps users quickly understand the scope and link to the documentation (right side bar on the repository page):

- Publishing seems to be manual right now and don't understand how that is setup? I usually use pyscaffold and it makes it easier for some routine package maintenance tasks.

- code quality (linting) is not included in the CI/CD tests.

- As I was reviewing the package, I thought a couple of features may be super helpful:
  - Highlight paths/features of interest
  - Ability to choose or specify branch widths

Overall, the package provides an intuitive and easy way to integrate phylogeny visualizations with matplotlib based workflows.
