### 2024-08-06 fd799c2
removed decimal from branch supports which are 1
