### 2024-08-12 #160 @ctb
@jkanche https://pyscaffold.org/en/stable/ looks fascinating!
