### 2024-08-22 df55802
initial commit for collapse function
