### 2024-08-27 #160 @ctb
@jkanche and @ammaraziz, thank you so much for your reviews!!

@KatyBrown, I've extracted the unchecked bits of the reviews below -- please see bottom for my thoughts!

# pyopensci -- summary of reviews by CTB

## Package review by @jkanche:

In general please consider whether:

---

- Cannot run tests following the instructions

I cloned the repo and installed it in edit mode:

(errors elided)

- Add a short description and link to docs for the repository metadata. Helps users quickly understand the scope and link to the documentation (right side bar on the repository page):

- Publishing seems to be manual right now and don't understand how that is setup? I usually use pyscaffold and it makes it easier for some routine package maintenance tasks.

- code quality (linting) is not included in the CI/CD tests.

- As I was reviewing the package, I thought a couple of features may be super helpful:
  - Highlight paths/features of interest
  - Ability to choose or specify branch widths

**CTB comment**: you could put these in the issue tracker, if you didn't want to implement them now :)

Overall, the package provides an intuitive and easy way to integrate phylogeny visualizations with matplotlib based workflows. Let me know if you have any questions.

## Package Review by @ammaraziz

---

Great package, works nicely and the provided example is a great start. The power of matplotlib means it's possible to make great plots.

1. Similar to the previous review, when testing the package I get errors. However, the documentation states:

> Run unit tests in the root directoy as `pytest --cov=plot_phylo` and ensure that they pass and that coverage is 100%.

But I get an error:
>pytest: error: unrecognized arguments: --cov=plot_phylo

Testing works if you run just `pytest` in the root dir:

(elided)

2. Consider adding some basic usage docs to the readme to get people started.
3. Provide example of how to color internal nodes and branches.
4. Consider using github actions for testing and publishing.
5. Add a section on how to cite your work, even if it's just the github repo, for example:
```
# Cite
KatyBrown - plot_phylo https://github.com/KatyBrown/plot_phylo
```

Good job on the great package.

---

## Editor comments from CTB

The three major suggestions from the reviews that stand out are -

* a few more usage docs are suggested/would be welcome!
* test the instructions for running the tests, and/or improve the documentation;
* suggest using GitHub CI - lmk if you need some pointers to quickstarts! I have some projects that do basically what you need, I think. Note that putting CI in place will help with the test running & documentation, above, too!

The rest all seem like fairly quick add-this/adjust-that changes.

Please let me know if you have thoughts or questions!
