### 2024-08-30 68625a0
added option to collapse nodes
