### 2024-09-30 7d40980
moving functions after drawing the tree to the post_draw module file
