+++ plot_phylo/post_draw.py
@@ -1,0 +1,210 @@
+#!/usr/bin/env python3
+def get_boxes(ax, texts):
+    '''
+    Converts a list of text objects to their co-ordinates on the axis in
+    axis units.
+
+    Parameters
+    ----------
+    ax : matplotlib.axes._axes.Axes,
+        An open matplotlib ax object where the tree will be plotted.
+    texts : list
+        A list of matplotlib.pyplot.text objects representing the tip
+        labels from the tree and their metadata.
+
+    Returns
+    -------
+    boxpos : dict
+        A dictionary of dictionaries where the top level keys are the tip
+        labels. Within each subdictionary the key value pairs are
+        index (position in the tree starting from the top) and the minimum,
+        maximum and central position of the text box on the x and y axis,
+        as index, xmin, xmid, xmax and ymin, ymid and ymax.
+
+    '''
+    boxpos = dict()
+    # Iterate through the tip labels
+    for i, txt in enumerate(texts):
+        # Get the position of the boxes containing the labels, convert to axis
+        # units
+        box = ax.transData.inverted().transform(txt.get_window_extent(
+            ax.figure.canvas.get_renderer()))
+        nam = txt.get_text().strip()
+        # Build the dictionary
+        boxpos[nam] = dict()
+        boxpos[nam]['index'] = i
+        boxpos[nam]['xmin'] = box[0][0].round(3)
+        boxpos[nam]['xmax'] = box[1][0].round(3)
+        boxpos[nam]['ymin'] = box[0][1].round(3)
+        boxpos[nam]['ymax'] = box[1][1].round(3)
+        boxpos[nam]['ymid'] = (box[0][1] + ((
+            box[1][1] - box[0][1]) / 2)).round(3)
+        boxpos[nam]['xmid'] = (
+            box[0][0] + ((box[1][0] - box[0][1]) / 2)).round(3)
+
+    return (boxpos)
+
+
+def auto_axis(subplot, textobj, xpos, ypos, width, height, depth, scale_bar,
+              branch_lengths):
+
+    xint = width * 0.1
+    yint = height / depth[2]
+
+    ymin = ypos - (yint * 0.5)
+    ymax = ypos + height + (yint * 0.5)
+    if scale_bar:
+        ymin -= yint + (yint * 0.2)
+    subplot.set_ylim(ymin, ymax)
+    xmin = xpos - xint
+    nboxes = get_boxes(subplot, textobj)
+    xmaxes = [nboxes[x]['xmax'] for x in nboxes]
+    xmax = xint + max(xmaxes)
+    subplot.set_xlim(xmin, xmax)
+    return (textobj, subplot)
+
+def draw_scale_bar(ax, width, height, depth, left, bottom,
+                   scale_bar_width=None,
+                   appearance={'font_size': 10}):
+    '''
+    Adds a scale bar to the tree - only when branch lengths are specified.
+
+    The default length is 0.25 * total tree width, or it can be specified
+    using scale_bar_width.
+
+    Parameters
+    ----------
+    ax : matplotlib.axes._axes.Axes
+        An open matplotlib ax object
+
+    height : float
+        Height of the tree, in axis units.
+    width : float
+        Width of the tree, in axis units.
+    depth: tuple(float, float, float)
+        Total height and width of the original tree in terms of number of
+        nodes, total branch length, number of tips
+    bottom: float
+        The bottom position of the tree on the y axis
+    scale_bar_width: float
+        The desired width of the scale bar, if not specified
+        tree width * 0.25 is used
+    appearance: dict
+        Dictionary of parameters specifying the appearance of the tree.
+    '''
+    # depth[1] - total width of the tree in tree units
+    # width - total width of tree in axis units
+    # xint - width of one tree unit in axis units
+    xint = width / depth[1]
+
[…]
