### 2024-09-30 21ad6d4
documenting auto ax function
