### 2024-09-30 b44a95e
Revert "initial commit for collapse function"
This reverts commit df55802aded328d9e16018e4800423932beba836.
