### 2024-09-30 8e83446
Revert "removed decimal from branch supports which are 1"
This reverts commit fd799c24d84b579dd7eb185c2864eeaa4fdfdc12.
