### 2024-10-01 46ed6ad
edited testing instructions in CONTRIBUTING.md
