### 2024-10-01 a332ec8
updated testing instructions in CONTRIBUTING.md (again)
