### 2024-10-01 c694905
updated manual to include testing
