+++ man/contents.rst
@@ -6,6 +6,7 @@ plot_phylo
 
    index.md
    pages/installation.md
+   pages/testing.md
    pages/quickstart.md
    pages/parameters.md
    pages/matplotlib.md
