### 2024-10-01 390d46c
updated pyproject.toml with metadata
