+++ pyproject.toml
@@ -4,3 +4,34 @@ requires = [
     "wheel"
 ]
 build-backend = "setuptools.build_meta"
+
+[project]
+name = "plot-phylo"
+version = "0.1.10"
+description = "This module allows the user to plot a phylogenetic tree on an existing matplotlib axis."
+readme = "README.md"
+license = { text = "MIT" }
+keywords = ["phylogenetics", "matplotlib"]
+
+# Author information
+authors = [
+    { name = "Katy Brown", email = "kab84@cam.ac.uk" }
+]
+
+# Maintainer information (optional)
+maintainers = [
+    { name = "Katy Brown", email = "kab84@cam.ac.uk" }
+]
+
+# Project URLs (optional but helpful for linking to docs, source, etc.)
+urls = {
+    "Documentation" = "https://plot-phylo.readthedocs.io",
+    "Source Code" = "https://github.com/KatyBrown/plot_phylo",
+    "Bug Tracker" = "https://github.com/KatyBrown/plot_phylo/issues"
+}
+
+dependencies = [
+    "ete3 >= 3.1.0",
+    "matplotlib >= 2.1.1"
+    "python >= 3.6"
+]
