### 2024-10-01 5a0a41b
fixed typo in pyproject.toml
