+++ pyproject.toml
@@ -32,6 +32,6 @@ urls = {
 
 dependencies = [
     "ete3 >= 3.1.0",
-    "matplotlib >= 2.1.1"
+    "matplotlib >= 2.1.1",
     "python >= 3.6"
 ]
