+++ pyproject.toml
@@ -30,8 +30,12 @@ urls = {
     "Bug Tracker" = "https://github.com/KatyBrown/plot_phylo/issues"
 }
 
+# Correct placement of dependencies
 dependencies = [
     "ete3 >= 3.1.0",
-    "matplotlib >= 2.1.1",
-    "python >= 3.6"
+    "matplotlib >= 2.1.1"
 ]
+
+# Python version requirements
+[project.requires-python]
+version = ">=3.6"
