### 2024-10-01 45b560f
added additional info to README.md
