+++ plot_phylo/__init__.py
@@ -1,3 +1,1 @@
-from plot_phylo.plot_phylo import *
-
 __author__ = 'Katy Brown, Duncan Cross'
