+++ plot_phylo/__init__.py
@@ -1,1 +1,2 @@
+from plot_phylo import *
 __author__ = 'Katy Brown, Duncan Cross'
