+++ plot_phylo/__init__.py
@@ -1,2 +1,3 @@
-from plot_phylo import *
+from plot_phylo.plot_phylo import *
+
 __author__ = 'Katy Brown, Duncan Cross'
