### 2024-10-01 1d0c743
added quickstart info to README.md
