### 2024-10-01 b0b1623
added details of CI to README.md
