### 2024-10-01 ebf13f0
add CI info to CONTRIBUTING.md
