### 2024-10-01 KatyBrown/plot_phylo#6 @KatyBrown
add ability to highlight nodes / paths of interest automatically
as discussed in the PyOpenSci review [here](https://github.com/pyOpenSci/software-submission/issues/160)
