### 2024-10-01 KatyBrown/plot_phylo#7 @KatyBrown
add option to specify / change branch widths 
as discussed in the PyOpenSci review [here](https://github.com/pyOpenSci/software-submission/issues/160)
