### 2024-10-01 #160 @KatyBrown
Hi everyone, apologies for my extremely slow response to this. I really appreciate all the feedback on my package.
I've addressed specific comments below.
