### 2024-10-01 f9f6144
edited automatic pip release workflow
