### 2024-10-01 #160 @KatyBrown
@jkanche

> #### Review Comments
>
> 	* Cannot run tests following the instructions

Sorry about that, I’ve [added new instructions to the manual](https://plot-phylo.readthedocs.io/en/latest/pages/testing.html) - which hopefully now work.

> 	* Add a short description and link to docs for the repository metadata. Helps users quickly understand the scope and link to the documentation (right side bar on the repository page):

I have now added these to the main page.

> 	* Publishing seems to be manual right now and don't understand how that is setup? I usually use pyscaffold and it makes it easier for some routine package maintenance tasks.

I’ve now added publishing automatically to pip as a [GitHub workflow](https://github.com/KatyBrown/plot_phylo/blob/main/.github/workflows/python-publish.yml) - I’ll also look into this package as it sounds useful!

>
> 	* code quality (linting) is not included in the CI/CD tests.

It is [now included](https://github.com/KatyBrown/plot_phylo/blob/main/.github/workflows/main.yml)

> 	* As I was reviewing the package, I thought a couple of features may be super helpful:

>   	* Highlight paths/features of interest
>   	* Ability to choose or specify branch widths

> **CTB comment**: you could put these in the issue tracker, if you didn't want to implement them now :)

I’ve added these two as enhancements in the [issue tracker](https://github.com/KatyBrown/plot_phylo/issues) - as I have a few other enhancements I’d like to make at the same time and I don't want to delay this even further.

>
> Overall, the package provides an intuitive and easy way to integrate phylogeny visualizations with matplotlib based workflows. Let me know if you have any questions.

Thank you very much!
