### 2024-10-01 #160 @KatyBrown
@ammaraziz
>
> Great package, works nicely and the provided example is a great start. The power of matplotlib means it's possible to make great plots.

Thank you! I appreciate it and I’m glad it works!

> 	1. Similar to the previous review, when testing the package I get errors. However, the documentation states:
>
> > Run unit tests in the root directoy as `pytest --cov=plot_phylo` and ensure that they pass and that coverage is 100%.
>
> But I get an error:
>
> > pytest: error: unrecognized arguments: --cov=plot_phylo
>
> Testing works if you run just `pytest` in the root dir:

Apologies for this, I’ve  [updated the manual](https://plot-phylo.readthedocs.io/en/latest/pages/testing.html) - with instructions which hopefully now work.

> 	2. Consider adding some basic usage docs to the readme to get people started.

I’ve added a QuickStart guide to the [README.md](https://github.com/KatyBrown/plot_phylo/blob/main/README.md) along with an additional link to the full documentation, so it’s easier to find.

> 	3. Provide example of how to color internal nodes and branches.

I already have this example [here](https://plot-phylo.readthedocs.io/en/latest/pages/parameters.html#col-dict) but I’ve tried to make the manual easier to find from the README.

> 	4. Consider using github actions for testing and publishing.

I was already using GitHub actions for CI testing [here](https://github.com/KatyBrown/plot_phylo/actions), I’ve added details of this to the [README.md](https://github.com/KatyBrown/plot_phylo/blob/main/README.md)  and [CONTRIBUTING.md](https://github.com/KatyBrown/plot_phylo/blob/main/CONTRIBUTING.md) files so it is easier to find. I’ve also now added publishing automatically to pip as a [GitHub workflow](https://github.com/KatyBrown/plot_phylo/actions/workflows/python-publish.yml)

> 	5. Add a section on how to cite your work, even if it's just the github repo, for example:

Thank you - I’ve added this now to the [README](https://github.com/KatyBrown/plot_phylo/blob/main/README.md)

> Good job on the great package.

Thank you very much!
