### 2024-10-01 #160 @KatyBrown
@ctb

> The three major suggestions from the reviews that stand out are -
>
> 	* a few more usage docs are suggested/would be welcome!

I think it was not very clear where to find my documentation, I’ve added additional links. If there is [still not enough detail](https://plot-phylo.readthedocs.io/en/latest/index.html) please let me know.

>
> 	* test the instructions for running the tests, and/or improve the documentation;

I’ve  [updated the manual](https://plot-phylo.readthedocs.io/en/latest/pages/testing.html) - with instructions which hopefully now work.

> 	* suggest using GitHub CI - lmk if you need some pointers to quickstarts! I have some projects that do basically what you need, I think. Note that putting CI in place will help with the test running & documentation, above, too!

Thank you. I was already using GitHub actions for CI testing [here](https://github.com/KatyBrown/plot_phylo/actions), I’ve added details of this to the [README.md](https://github.com/KatyBrown/plot_phylo/blob/main/README.md)  and [CONTRIBUTING.md](https://github.com/KatyBrown/plot_phylo/blob/main/CONTRIBUTING.md) files so it is easier to find. I’ve also now added publishing automatically to pip as a [GitHub workflow](https://github.com/KatyBrown/plot_phylo/actions/workflows/python-publish.yml)

> The rest all seem like fairly quick add-this/adjust-that changes.
>
> Please let me know if you have thoughts or questions!

Thank you!

@ctb I’ve also done the following to address additional things on the checklist:

* Added all the required badges to my README.md
* Updated my pyproject.toml to include metadata
* Added a little more information about how the project fits in amongst existing packages and why it is useful to the README.md.
