### 2024-11-27 a47c8a6
split functions into draw_tree module file
