+++ plot_phylo/draw_tree.py
@@ -1,0 +1,308 @@
+import numpy as np
+
+
+def collapse_nodes(tree, collapse_list, collapse_names):
+    cD = dict(zip(collapse_list, collapse_names))
+    collapseD = dict()
+    for string in collapse_list:
+        keeps = set()
+        collapsed = set()
+        done = set()
+        ddD = dict()
+        for node in tree.traverse():
+            x = 0
+            L = list(node.get_leaves())
+            dd = []
+            for leaf in L:
+                if leaf.name.endswith(string) and leaf not in done:
+                    dd.append(leaf.dist)
+                    x += 1
+            if x == len(L) or (len(L) == 1 and leaf not in done):
+                keeps.add(L[0].name)
+                done = done | set(L)
+                if x > 1:
+                    collapsed.add(L[0])
+                    ddD[L[0]] = np.mean(dd)
+        tree.prune(keeps)
+        for leaf in tree.get_leaves():
+            if leaf in collapsed:
+                leaf.dist = ddD[leaf]
+                leaf.name = 'COLLAPSE|%s' % (leaf.name)
+                collapseD[leaf.name] = cD[string]
+    return (tree, collapseD)
+
+
+def draw_tree(tree, ax,
+              x=0,
+              y=0,
+              x0=0,
+              ps=[],
+              height=10,
+              width=10,
+              depth=None,
+              align_tips=False,
+              rev_align_tips=False,
+              branch_lengths=True,
+              reverse=False,
+              appearance={},
+              collapse=[],
+              collapseD=dict()):
+    '''
+    Plot a phylogenetic tree in matplotlib
+
+    Parameters
+    ----------
+    tree : ete3.Tree
+        ete3 Tree object
+    ax : matplotlib.axes._axes.Axes
+        An open matplotlib ax object
+    x : float
+        Current position on x axis.
+    y : float
+        Current position on y axis.
+    ps: list
+        Used internally, list of tip labels and x and y positions
+    height: float
+        Height of the tree in axis units
+    width: float
+        Width of the tree in axis units
+    depth: tuple(float, float, float)
+        Total height and width of the original tree in terms of number of
+        nodes, total branch length, number of tips
+    align_tips: bool
+        If True, the tip labels will be aligned rather than positioned at
+        the end of the branches. Default False.
+    rev_align_tips: bool
+        If True  the tip labels are right aligned
+        if reverse=False and left aligned if reverse=True.
+    branch_lengths: bool
+        If True, use the branch lengths provided in the tree, otherwise fix
+        all branches to the same length. Default True.
+    reverse: bool
+        If True, reverse the tree on the y-axis, showing the root on the right
+        hand side. Default False.
+    appearance: dict
+        Dictionary of parameters specifying the appearance of the tree.
+    collapse: list
+        Collapse nodes where possible based on strings in the list.
+
+    Returns
+    -------
+    y   float
+        Position of previous node on y axis.
+    ym  float
+        Position of current node on y axis.
+    ps  list
+        List of lists - ordered as tip labels, tip label text objects,
+        alignment lines (if aligned). All are in the same order.
+    '''
+    # This is the increment for the position of each terminal node on
+    # the y axis.
+    # The number of nodes - 1 is used because one branch will be at position 0
+    yint = height / (depth[2] - 1)
+
+    # td is the branch length - if the branch_lengths parameter is False
+    # it is set to 1
+
+    # If the branch lengths are used, the total width of the tree is in
+    # the tree branch units, otherwise it's the total number of nodes from
+    # the root to the farthest branch
+
+    # textinc stops the tip labels from being immediately next to the tips
+
+    if branch_lengths:
+        td = tree.dist
+        tot_width = depth[1]
+
+    else:
+        td = 1
+        tot_width = depth[0] + 1
+
[…]
