+++ setup.py
@@ -24,9 +24,7 @@ setuptools.setup(
      long_description_content_type="text/markdown",
      url="https://github.com/KatyBrown/plot_phylo",
      packages=setuptools.find_packages(),
-     package_dir={'plot_phylo':'plot_phylo'},
      install_requires=['matplotlib', 'ete3'],
-     scripts=['plot_phylo/plot_phylo.py'],
      classifiers=[
          "Programming Language :: Python :: 3",
          "License :: OSI Approved :: MIT License",
