+++ README.md
@@ -37,7 +37,7 @@ You can also download the latest release [here](https://github.com/KatyBrown/plo
 
 Or clone the GitHub repository directly.
 
-`git clone git@github.com:KatyBrown/plot_phylo.git`
+`git clone https://github.com/KatyBrown/plot_phylo.git`
 
 ## Quick Start
 For detailed usage instructions, visit our [ReadTheDocs page](https://plot-phylo.readthedocs.io/en/latest/index.html).
