+++ README.md
@@ -2,7 +2,7 @@
 [![Documentation Status](https://readthedocs.org/projects/plot-phylo/badge/?version=latest)](https://plot-phylo.readthedocs.io/en/latest/?badge=latest)<br>
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)<br>
 ![Python Versions](https://img.shields.io/pypi/pyversions/plot-phylo)<br>
-![PyPI - Version](https://img.shields.io/pypi/v/plot-phylo)
+[![PyPI - Version](https://img.shields.io/pypi/v/plot-phylo)](https://pypi.org/project/plot-phylo/)
 
 # plot_phylo
 
