### 2024-12-30 #160 @ctb
Hi @KatyBrown apologies for the delay, busy December! I submitted a small PR to fix a few things in the README (see https://github.com/KatyBrown/plot_phylo/pull/8) but it all looks good to me!

If @ammaraziz and @jkanche could update their checkboxes that would be ideal, but I've verified most of them myself.

I'll figure out what needs to be done to fully accept when I'm back in a week. Happy New Year!
