### 2025-01-02 #160 @jkanche
This looks good to me. Optionally, you may also create a citation.cff file in the repo. GitHub add a new menu item on the right sidebar to quickly copy the citation. But otherwise looks good. Thank you for all the work and Happy new year!
