### 2025-01-07 a35a182
fixed new compatibility issue between python 3.7 and latest ubuntu in main.yml workflow
