+++ .github/workflows/main.yml
@@ -12,17 +12,20 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, macos-latest, windows-latest ]
-        python-version: [ 3.7, 3.8, 3.9, '3.10', 3.11, 3.12 ]
-        exclude:  # Apple Silicon ARM64 does not support Python < v3.8
+        python-version: [ 3.8, 3.9, '3.10', 3.11, 3.12 ]
+        include:  # python 7 sometimes requires legacy versions
           - python-version: "3.6"
-            os: macos-latest
-          - python-version: "3.7"
-            os: macos-latest
-        include:  # So run those legacy versions on Intel CPUs
+            os:  "macos-13"
           - python-version: "3.6"
-            os: macos-13
+            os: "ubuntu-22.04"
+          - python-version: "3.6"
+            os: "windows-latest"
           - python-version: "3.7"
-            os: macos-13
+            os:  "macos-13"
+          - python-version: "3.7"
+            os: "ubuntu-22.04"
+          - python-version: "3.7"
+            os: "windows-latest"
       fail-fast: false
     # a job is a seq of steps
     steps:
