### 2025-01-07 e8c5e16
tried again to fix ubuntu and python 3.6 compatibility in main.yml workflow
