### 2025-01-07 98d2bf9
removed unsupported ubuntu version from CI workflow
