### 2025-01-07 #160 @KatyBrown
> This looks good to me. Optionally, you may also create a citation.cff file in the repo. GitHub add a new menu item on the right sidebar to quickly copy the citation. But otherwise looks good. Thank you for all the work and Happy new year!

Thanks - I've added the citation.cff file.
