### 2025-01-21 #160 @ctb
---

🎉 plot-phylo has been approved by pyOpenSci! Thank you @KatyBrown  for submitting <package-name> and many thanks to @jkanche @ammaraziz for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
