### 2025-01-21 #160 @ctb
dear @jkanche @ammaraziz @KatyBrown if you could fill out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7) (it's very brief!) we would appreciate it!

also, please let me know if you would like to join the pyopensci slack, and if so drop me an e-mail with your preferred invite e-mail at titus@idyll.org.

last but not least, @KatyBrown - we wanted to invite you to write a blog post (totally optional) on your package for us to promote your work! Here are a few examples of other blog posts:

[pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation)
[movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis)

and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post.

You could write a tutorial, too!

If you are too busy for this no worries. But if you have time - we'd love to spread the word about your package!
