### 2025-02-21 #160 @lwasser
Hi friends things seem to be moving forward here! i'm just slowly helping to wrap up reviews that are winding down!

@KatyBrown
can you please provide the following so we can wrap things up? I know times are challenging and I really appreciate you energy working on this! The blog post or tutorial (or both?) could be done at any time! but to close this review we need the following:

First, we need an achive of the version accepted of your package along with the version accepted so we can update the header of this issue! Zenodo is our preferred archive and there is an overview [here](https://www.pyopensci.org/lessons/package-share-code/publish-share-code/cite-code.html#cite-your-code) of using it. Share any questions here!

Add the pyopensci badge to your readme file at the top.
- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

Please let us know if you have any questions! Hopefully the above tasks as quick for you to do!
