+++ README.md
@@ -3,6 +3,8 @@
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)<br>
 ![Python Versions](https://img.shields.io/pypi/pyversions/plot-phylo)<br>
 ![PyPI - Version](https://img.shields.io/pypi/v/plot-phylo)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.13871592.svg)](https://doi.org/10.5281/zenodo.13871592)
+[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)
 
 # plot_phylo
 
