### 2025-03-07 #160 @KatyBrown
@lwasser thank you!  I am working on the blog post etc but keep getting distracted, sorry, I am always slow. I'm glad we can wrap things up without those being finished.

- I already have Zenodo set up, the DOI is 10.5281/zenodo.13871592, I've now added the DOI link to the README.
- I've also added the pyOpenSci badge, thank you.
- I think I already did the post review survey.

Thanks again,
Katy
