### 2025-03-19 #160 @lwasser
@KatyBrown no worries at all!! Please don't apologize; I'm just wrapping up loose ends!! I've updated the header of this issue. I think we are now good to close this review as completed!! Plot_Phylo will have a zenodo badge on it in our listing [here](https://www.pyopensci.org/python-packages.html) when our next cron job runs and updates things 🚀

If you even want to connect with us directly, we can also invite you (and anyone involved in this review to slack) at any time!!

Feel free to ping us here if you have any questions or challenges with the blog post. Again it's not required, but it does help us promote your package! Closing this now!!
